+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.0] - 2025-12-15
+
 ### Added
 - Create Response classes and ViewConfig infrastructure for Responder pattern [SCI-30]
   - Created Response base classes: `ResponseInterface` and `AbstractResponse`